requirements.txt
asgiref==3.2.7
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
defusedxml==0.6.0
dj-rest-auth==1.0.4
Django==3.0.5
django-allauth==0.41.0
django-backblazeb2-storage==0.4
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-phone-verify==1.1.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
face-recognition-models==0.3.0
future==0.18.2
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.9
Markdown==3.2.1
nexmo==2.4.0
numpy==1.18.2
oauthlib==3.1.0
phonenumbers==8.12.1
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.12.0
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2019.3
razorpay==1.2.0
requests==2.23.0
requests-oauthlib==1.3.0
rest-social-auth==4.1.0
rsa==4.0
sendgrid==6.2.2
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
sqlparse==0.3.1
twilio==6.38.0
uritemplate==3.0.1
urllib3==1.25.8
